Move the four TUNE screens out of EcuLab.jsx
TuningGrid and SelectionDock are used by three of the four TUNE views (AIR,
SPARK, FUEL — ECU uses neither), so they move to src/ui/components/ rather
than into any one screen, joining PickList there for the same reason. Both
data-testid attributes (tuning-grid, selection-dock) survive the move
unchanged.

TUNE is a mutually-exclusive conditional render, not an accordion like BUILD
or DASH: exactly one of the four screens is in the document at a time, and
`goSection`/the raw tab-strip switcher stay in the shell, not
`makeToggleSection`. The switcher itself is left raw per issue #83's call.

gridProps (the `{ selection, setSelection }` spread EcuLab.jsx built once and
passed to every TuningGrid) does not survive extraction: each screen now
reads `selection` off the store itself via useTune() and defines its own
one-line `setSelection`, so there is nothing left to spread.

Multi-consumer derivations (veAdvice, veTruth, calAdvice, dutyPreview, fuel,
injectorCc, needsMafRecal, chartData, result) stay in the shell and are
passed down. `recalcVE` and `dutyDangerous` had exactly one reader each
(VeScreen and EcuScreen respectively) and move with their screens, computed
off the shell-owned values they were passed.

Markup is unchanged from EcuLab.jsx — inline styles carry over verbatim.
Stylesheets land in the next commit.

Moving TUNE's ECU-Injector Seg call site out from under EcuLab.jsx's
@ts-nocheck surfaced that Seg's own JSDoc types had only ever promised
`id: string`, when two call sites (this one and DYNO's manifold-pressure
picker) always keyed options off a number. Widened the type to
`string|number`, which is what the component already did at runtime, and
added a test proving it.

button-call-sites.test.jsx's source scan now also walks src/ui/components/:
SelectionDock's DONE button moved there, and the glob previously only
covered EcuLab.jsx and src/ui/screens/. The real total is unchanged at 23 —
three call sites moved, none were added or deleted — so the floor stays put.
